requirement.txt
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
